split cpp
